refactor: Column -> DDLColumn 네이밍 수정 및 공통 메서드 추출
